Moving timing and the frame limiter to core. Moving circular buffer and fps counter to a new utils crate
