+++ Cargo.toml
@@ -27,6 +27,7 @@ amethyst_config = { path = "amethyst_config", version = "0.4.0" }
 amethyst_core = { path = "amethyst_core", version = "0.1.0" }
 amethyst_renderer = { path = "amethyst_renderer", version = "0.5" }
 amethyst_input = { path = "amethyst_input", version = "0.1" }
+amethyst_utils = { path = "amethyst_utils", version = "0.1" }
 cgmath = "0.14"
 cpal = "0.4"
 crossbeam = "0.3"
@@ -87,4 +88,4 @@ name = "sphere_separate"
 path = "examples/07_separate_sphere/main.rs"
 
 [workspace]
-members = ["amethyst_assets", "amethyst_config", "amethyst_core", "amethyst_input", "amethyst_renderer"]
+members = ["amethyst_assets", "amethyst_config", "amethyst_core", "amethyst_input", "amethyst_renderer", "amethyst_utils"]
