+++ amethyst_core/src/frame_limiter.rs
