+++ amethyst_core/src/lib.rs
@@ -1,8 +1,14 @@
 extern crate cgmath;
 extern crate fnv;
+#[macro_use]
+extern crate serde;
 extern crate specs;
 
 pub use bundle::ECSBundle;
+pub use transform::*;
+pub use timing::*;
 
 pub mod bundle;
 pub mod transform;
+pub mod timing;
+pub mod frame_limiter;
