+++ amethyst_core/src/timing.rs
