+++ amethyst_utils/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "amethyst_core"
+name = "amethyst_utils"
 version = "0.1.0"
 authors = ["Simon Rönnberg <seamonr@gmail.com>"]
-description = "Amethyst core"
+description = "Amethyst utils"
 
-documentation = "https://www.amethyst.rs/doc/master/amethyst_core/"
+documentation = "https://www.amethyst.rs/doc/master/amethyst_utils/"
 homepage = "https://www.amethyst.rs/"
 repository = "https://github.com/amethyst/amethyst"
 
@@ -15,7 +15,5 @@ appveyor = { repository = "amethyst/amethyst" }
 travis-ci = { repository = "amethyst/amethyst" }
 
 [dependencies]
+amethyst_core = { path = "../amethyst_core", version = "0.1.0" }
 specs = "0.10"
-fnv = "1.0"
-cgmath = "0.14"
-serde = { version = "1", features = ["serde_derive"] }
