+++ amethyst_utils/src/circular_buffer.rs
