+++ amethyst_utils/src/lib.rs
@@ -1,0 +1,5 @@
+extern crate specs;
+extern crate amethyst_core;
+
+pub mod fps_counter;
+pub mod circular_buffer;
