+++ examples/04_pong/systems/move_balls.rs
@@ -1,7 +1,7 @@
 use Ball;
 use amethyst::core::transform::LocalTransform;
 use amethyst::ecs::{Fetch, Join, System, WriteStorage};
-use amethyst::timing::Time;
+use amethyst::core::timing::Time;
 
 /// This system is responsible for moving all balls according to their speed
 /// and the time passed.
