+++ examples/04_pong/systems/paddle.rs
@@ -2,7 +2,7 @@ use Paddle;
 use amethyst::core::transform::LocalTransform;
 use amethyst::ecs::{Fetch, Join, System, WriteStorage};
 use amethyst::input::InputHandler;
-use amethyst::timing::Time;
+use amethyst::core::timing::Time;
 
 /// This system is responsible for moving all the paddles according to the user
 /// provided input.
