+++ src/ecs/mod.rs
@@ -8,4 +8,3 @@ pub use specs::*;
 
 pub mod audio;
 pub mod input;
-pub mod util;
