--- src/ecs/util/mod.rs
@@ -1,3 +1,0 @@
-//! Utilities that help interact with the ecs system.
-pub mod systems;
-pub mod resources;
