+++ src/lib.rs
@@ -60,6 +60,7 @@ pub extern crate amethyst_config as config;
 pub extern crate amethyst_core as core;
 pub extern crate amethyst_input as input;
 pub extern crate amethyst_renderer as renderer;
+pub extern crate amethyst_utils as utils;
 pub extern crate shrev;
 pub extern crate winit;
 
@@ -77,9 +78,6 @@ extern crate shred;
 extern crate smallvec;
 extern crate specs;
 
-#[macro_use]
-extern crate serde_derive;
-
 pub use self::app::{Application, ApplicationBuilder};
 pub use self::engine::Engine;
 pub use self::error::{Error, Result};
@@ -89,8 +87,6 @@ pub mod audio;
 pub mod ecs;
 pub mod event;
 pub mod prelude;
-pub mod timing;
-pub mod util;
 
 mod app;
 mod engine;
