--- src/util/mod.rs
@@ -1,4 +1,0 @@
-//! Util data structures and functions for various common tasks
-pub mod time;
-pub mod circular_buffer;
-pub mod frame_limiter;
